% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{BayesHMM-package}
\alias{BayesHMM}
\alias{BayesHMM-package}
\title{BayesHMM: Full Bayesian Inference for Hidden Markov Models}
\description{
An R Package to run full Bayesian inference on Hidden Markov Models (HMM) using the probabilistic programming language Stan. The software enables users to fit HMM with time-homogeneous transitions as well as time-varying transition probabilities. Priors can be set for every model parameter. Implemented inference algorithms include forward (filtering), forward-backwards (smoothing), Viterbi (most likely hidden path), prior predictive sampling, and posterior predictive sampling. Graphs, tables and other convenience methods for convergence diagnosis, goodness of fit, and data analysis are provided.
}
\details{
See the Introduction vignette: vignette("introduction", package = "BayesHMM")
Additionally, you may start with the manual help for ?specify and ?fit.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/luisdamiano/BayesHMM/}
  \item \url{https://summerofcode.withgoogle.com/projects/#4681157036212224}
  \item Report bugs at \url{https://github.com/luisdamiano/BayesHMM/issues}
}

}
\author{
\strong{Maintainer}: Luis Damiano \email{damiano.luis@gmail.com} (0000-0001-9107-0706) [copyright holder]

Authors:
\itemize{
  \item Michael Weylandt \email{michael.weylandt@gmail.com}
  \item Brian Peterson \email{brian@braverock.com}
}

}
\keyword{internal}
