% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityBeta.R
\name{Beta}
\alias{Beta}
\title{Beta density (univariate, continuous, bounded space)}
\usage{
Beta(alpha = NULL, beta = NULL, bounds = list(NULL, NULL),
  trunc = list(NULL, NULL), k = NULL, r = NULL, param = NULL)
}
\arguments{
\item{alpha}{Either a fixed value or a prior density for the first shape parameter.}

\item{beta}{Either a fixed value or a prior density for the second shape parameter.}

\item{bounds}{(optional) A list with two elements specifying the lower and upper bound for the parameter space. Use either a fixed value for a finite bound or NULL for no bounds. It defaults to an unbounded parameter space.}

\item{trunc}{(optional) A list with two elements specifying the lower and upper bound for the domain of the density function. Use either a fixed value for a finite bound or NULL for no truncation. It defaults to an unbounded domain.}

\item{k}{(optional) The number of the hidden state for which this density should be used. This argument is mostly for internal use: you should not use it unless you are acquainted with the internals of this software.}

\item{r}{(optional) The dimension of the observation vector dimension for which this density should be used. This argument is mostly for internal use: you should not use it unless you are acquainted with the internals of this software.}

\item{param}{(optional) The name of the parameter. This argument is mostly for internal use: you should not use it unless you are acquainted with the internals of this software.}
}
\value{
A \code{\link{Density}} object.
}
\description{
It can be used to specify either a
prior distribution for a model parameter or a likelihood function for an
observation model.
}
\examples{
# With fixed values for the parameters
Beta(1, 1)

# With priors for the parameters
Beta(
  alpha = Exponential(1), beta = Exponential(1)
)
}
\seealso{
Other Density: \code{\link{Bernoulli}},
  \code{\link{Binomial}}, \code{\link{Categorical}},
  \code{\link{Cauchy}}, \code{\link{CholeskyLKJCor}},
  \code{\link{Density}}, \code{\link{Dirichlet}},
  \code{\link{Exponential}}, \code{\link{GammaDensity}},
  \code{\link{Gaussian}}, \code{\link{ImproperUniform}},
  \code{\link{InitialSoftmax}},
  \code{\link{InverseWishart}},
  \code{\link{MVGaussianCholeskyCor}},
  \code{\link{MVGaussian}}, \code{\link{MVStudent}},
  \code{\link{Multinomial}},
  \code{\link{NegativeBinomialLocation}},
  \code{\link{NegativeBinomial}}, \code{\link{Poisson}},
  \code{\link{RegBernoulliLogit}},
  \code{\link{RegBinomialLogit}},
  \code{\link{RegBinomialProbit}},
  \code{\link{RegCategoricalSoftmax}},
  \code{\link{RegGaussian}}, \code{\link{Student}},
  \code{\link{TransitionFixed}},
  \code{\link{TransitionSoftmax}}, \code{\link{Wishart}}
}
\concept{Density}
