% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityBinomial.R
\name{Binomial}
\alias{Binomial}
\title{Binomial mass (univariate, discrete, bounded space)}
\usage{
Binomial(theta = NULL, N = NULL, ordered = NULL, equal = NULL,
  bounds = list(NULL, NULL), trunc = list(NULL, NULL), k = NULL,
  r = NULL, param = NULL)
}
\arguments{
\item{theta}{Either a fixed value or a prior density for the success proportion parameter.}

\item{N}{An integer with the number of trials (fixed quantity).}

\item{ordered}{(optional) A logical setting an increasing ordering constraint on any univariate parameter and any unconstrained parameter vector. Ordered simplices (e.g. \code{\link{Categorical}}, \code{\link{Multinomial}}, \code{\link{Dirichlet}}) are currently not implemented. Most useful for location parameters to break symmetries and fix label switching as shown in Betancourt (2017). It defaults to unordered parameters.}

\item{equal}{(optional) A logical setting whether the parameter takes the same value in every hidden state, i.e. the parameter is shared across states. It defaults to unequal parameters.}

\item{bounds}{(optional) A list with two elements specifying the lower and upper bound for the parameter space. Use either a fixed value for a finite bound or NULL for no bounds. It defaults to an unbounded parameter space.}

\item{trunc}{(optional) A list with two elements specifying the lower and upper bound for the domain of the density function. Use either a fixed value for a finite bound or NULL for no truncation. It defaults to an unbounded domain.}

\item{k}{(optional) The number of the hidden state for which this density should be used. This argument is mostly for internal use: you should not use it unless you are acquainted with the internals of this software.}

\item{r}{(optional) The dimension of the observation vector dimension for which this density should be used. This argument is mostly for internal use: you should not use it unless you are acquainted with the internals of this software.}

\item{param}{(optional) The name of the parameter. This argument is mostly for internal use: you should not use it unless you are acquainted with the internals of this software.}
}
\value{
A \code{\link{Density}} object.
}
\description{
It can be used to specify either a
prior distribution for a model parameter or a likelihood function for an
observation model.
}
\examples{
# With fixed values for the parameters
Binomial(0.5, 10)

# With priors for the parameters
Binomial(
  Beta(1, 1), 10
)
}
\references{
Betancourt, Michael (2017) Identifying Bayesian Mixture Models \emph{Stan Case Studies} \bold{Volume 4}. \href{https://mc-stan.org/users/documentation/case-studies/identifying_mixture_models.html}{Link}.
}
\seealso{
Other Density: \code{\link{Bernoulli}}, \code{\link{Beta}},
  \code{\link{Categorical}}, \code{\link{Cauchy}},
  \code{\link{CholeskyLKJCor}}, \code{\link{Density}},
  \code{\link{Dirichlet}}, \code{\link{Exponential}},
  \code{\link{GammaDensity}}, \code{\link{Gaussian}},
  \code{\link{ImproperUniform}},
  \code{\link{InitialFixed}}, \code{\link{InitialSoftmax}},
  \code{\link{InverseWishart}},
  \code{\link{MVGaussianCholeskyCor}},
  \code{\link{MVGaussian}}, \code{\link{MVStudent}},
  \code{\link{Multinomial}},
  \code{\link{NegativeBinomialLocation}},
  \code{\link{NegativeBinomial}}, \code{\link{Poisson}},
  \code{\link{RegBernoulliLogit}},
  \code{\link{RegBinomialLogit}},
  \code{\link{RegBinomialProbit}},
  \code{\link{RegCategoricalSoftmax}},
  \code{\link{RegGaussian}}, \code{\link{Student}},
  \code{\link{TransitionFixed}},
  \code{\link{TransitionSoftmax}}, \code{\link{Wishart}}
}
\concept{Density}
