% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityRegBernoulliLogit.R
\name{block_data.RegBernoulliLogit}
\alias{block_data.RegBernoulliLogit}
\title{Write the data block of the Stan code.}
\usage{
\method{block_data}{RegBernoulliLogit}(x, noLogLike)
}
\arguments{
\item{x}{A \code{\link{Density}} or a \code{\link{Specification}} object.}

\item{noLogLike}{A logical stating whether the log-likelihood should be excluded from the program. If TRUE, the Stan code will draw samples from the prior predictive density. If FALSE, the Stan code will draw samples from the posterior predictive density.}
}
\value{
A character vector with the Stan code.
}
\description{
Write the data block of the Stan code.
}
