% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificationInternals.R
\name{block_data.Specification}
\alias{block_data.Specification}
\title{Write the data block of the Stan code.}
\usage{
\method{block_data}{Specification}(x)
}
\arguments{
\item{x}{A \code{\link{Density}} or a \code{\link{Specification}} object.}
}
\value{
A character vector with the Stan code.
}
\description{
Write the data block of the Stan code.
}
