% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificationInternals.R
\name{block_functions}
\alias{block_functions}
\title{Write the function block of the Stan code.}
\usage{
block_functions(x)
}
\arguments{
\item{x}{A \code{\link{Density}} or a \code{\link{Specification}} object.}
}
\value{
A character string with the Stan code.
}
\description{
Write the function block of the Stan code.
}
\keyword{internal}
