% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificationInternals.R
\name{block_target.Specification}
\alias{block_target.Specification}
\title{Write the model block of the Stan code.}
\usage{
\method{block_target}{Specification}(x)
}
\arguments{
\item{x}{A \code{\link{Density}} or a \code{\link{Specification}} object.}
}
\value{
A character vector with the Stan code.
}
\description{
Write the model block of the Stan code.
}
