% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificationInternals.R
\name{block_tparameters.Specification}
\alias{block_tparameters.Specification}
\title{Write the transformed parameters block of the Stan code.}
\usage{
\method{block_tparameters}{Specification}(x)
}
\arguments{
\item{x}{A \code{\link{Density}} or a \code{\link{Specification}} object.}
}
\value{
A character string with the Stan code.
}
\description{
Write the transformed parameters block of the Stan code.
}
\keyword{internal}
