% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{browse_model}
\alias{browse_model}
\title{Load the underlying Stan code into an IDE or browser (platform dependent).}
\usage{
browse_model(fit, ...)
}
\arguments{
\item{fit}{An object returned by either \code{\link{sampling}} or \code{\link{optimizing}}.}

\item{...}{(optional) Arguments to be passed to \code{\link{browseURL}} (namely \emph{browser} and \emph{encodeIfNeeded}).}
}
\value{
No return value.
}
\description{
Load the underlying Stan code into an IDE or browser (platform dependent).
}
\seealso{
\code{\link{browseURL}}.
}
