% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{browse_model.stanfit}
\alias{browse_model.stanfit}
\title{Load the underlying Stan code into an IDE or browser.}
\usage{
\method{browse_model}{stanfit}(object, ...)
}
\arguments{
\item{object}{An object returned by \code{\link{specify}}, \code{\link{hmm}}, \code{\link{compile}}, \code{\link{draw_samples}}, \code{\link{fit}} or \code{\link{optimizing}}.}

\item{...}{(optional) Arguments to be passed to \code{\link{browseURL}} (namely \emph{browser} and \emph{encodeIfNeeded}).}
}
\value{
No return value.
}
\description{
The behavior is platform dependent.
}
\seealso{
\code{\link{browseURL}}.
}
\keyword{internal}
