% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{check}
\alias{check}
\title{Verify that the object is a valid specification. TO BE IMPLEMENTED.}
\usage{
check(spec)
}
\arguments{
\item{spec}{An object returned by either \code{\link{specify}} or \code{\link{hmm}}.}
}
\value{
A logical value with TRUE if the object is a valid specification or FALSE otherwise.
}
\description{
This function verifies that the structure of the object is valid. Useful to spot inconsistencies in the specification. \strong{TO BE IMPLEMENTED}.
}
