% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitIntegration.R, R/integration.R
\name{classify_alpha}
\alias{classify_alpha}
\title{Classify observations based on filtered probabilities.}
\usage{
classify_alpha(fit, ...)

classify_alpha(fit, reduce = NULL, chain = NULL)
}
\arguments{
\item{fit}{An object returned by either \code{\link{sampling}} or \code{\link{optimizing}}.}
}
\value{
A numeric vector with size equal to the time series length \emph{T} with values from 1 to the number of hidden states \emph{K}.

A numeric vector with size equal to the time series length \emph{T} with values from 1 to the number of hidden states \emph{K}.
}
\description{
This function assigns the hidden states at each time step \emph{t} to the hidden state \emph{k} with largest estimated filtered probability (\emph{alpha}).

This function assigns the hidden states at each time step \emph{t} to the hidden state \emph{k} with largest estimated filtered probability (\emph{alpha}).
}
