% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{classify_zstar}
\alias{classify_zstar}
\title{Assign the hidden states to the most likely path (\emph{zstar}).}
\usage{
classify_zstar(fit, reduce = NULL, chain = NULL)
}
\arguments{
\item{fit}{An object returned by either \code{\link{sampling}} or \code{\link{optimizing}}.}
}
\value{
A numeric vector with size equal to the time series length \emph{T} with values from 1 to the number of hidden states \emph{K}.
}
\description{
This is the result of Viterbi algorithm, which considers the joint distribution of the states.
}
