% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{compile}
\alias{compile}
\title{Compile a specified model.}
\usage{
compile(spec, priorPredictive = FALSE, writeDir = tempdir(), ...)
}
\arguments{
\item{spec}{An object returned by either \code{\link{specify}} or \code{\link{hmm}}.}

\item{...}{Arguments to be passed to rstan's \code{\link[rstan]{stan_model}}.}

\item{priorPredictive}{An optional logical stating whether the log-likelihood should be excluded from the program. If TRUE, the returned object can only be used to draw samples from the prior predictive density. If FALSE, the returned object can only be used to draw samples from the posterior predictive density. It defaults to FALSE.}

\item{writeDir}{An optional string with the path where the Stan file should be written. Useful to inspect and modify the Stan code manually. It defaults to a temporary directory.}
}
\value{
An instance of S4 class stanmodel.
}
\description{
This function turns the model specification into Stan code and compiles the program via rstan.
}
