% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityBinomial.R
\name{constants.Binomial}
\alias{constants.Binomial}
\title{Write the Stan code lines related with Density constants.}
\usage{
\method{constants}{Binomial}(x)
}
\arguments{
\item{x}{A \code{\link{Density}} object.}
}
\value{
A character string.
}
\description{
Write the Stan code lines related with Density constants.
}
\keyword{internal}
