% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityMultinomial.R
\name{constants.Multinomial}
\alias{constants.Multinomial}
\title{Write the Stan code lines related with Density constants.}
\usage{
\method{constants}{Multinomial}(x)
}
\arguments{
\item{x}{A \code{\link{Density}} object.}
}
\value{
A character vector.
}
\description{
Write the Stan code lines related with Density constants.
}
