% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{Density}
\alias{Density}
\title{Create a representation of a probability mass or density function. It can be
used to specify either a prior distribution for a model parameter or a
likelihood function for an observation model.}
\usage{
Density(name, ...)
}
\arguments{
\item{name}{TO BE DOCUMENTED}

\item{...}{TO BE DOCUMENTED}
}
\value{
TO BE DOCUMENTED
}
\description{
Description
}
\concept{Density}
