% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{DiscreteDensity}
\alias{DiscreteDensity}
\title{Create a representation of a probability mass or density function for a
discrete univariate random variable. It can be used to specify either a
prior distribution for a model parameter or a likelihood function for an
observation model.}
\usage{
DiscreteDensity(name, ...)
}
\arguments{
\item{name}{A character vector with the name of the density.}

\item{...}{Other arguments for the density.

#'}
}
\value{
#' @param x A \code{\link{Density}} object.
}
\description{
Create a representation of a probability mass or density function for a
discrete univariate random variable. It can be used to specify either a
prior distribution for a model parameter or a likelihood function for an
observation model.
}
\seealso{
Other Density: \code{\link{Binomial}},
  \code{\link{Categorical}}, \code{\link{InitialSoftmax}},
  \code{\link{RegBernoulliLogit}},
  \code{\link{RegBinomialLogit}},
  \code{\link{RegBinomialProbit}},
  \code{\link{RegCategoricalSoftmax}},
  \code{\link{RegGaussian}}, \code{\link{TransitionFixed}},
  \code{\link{TransitionSoftmax}}
}
\concept{DiscreteDensity}
