% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{drawSamples.Specification}
\alias{drawSamples.Specification}
\title{#'}
\usage{
drawSamples.Specification(spec, stanModel = NULL, y, x = NULL,
  u = NULL, v = NULL, writeDir = tempdir(), ...)
}
\description{
#'
}
