% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{explain}
\alias{explain}
\title{Create a user-friendly text describing the model.}
\usage{
explain(spec, observation = TRUE, initial = TRUE, transition = TRUE, print = TRUE)
}
\arguments{
\item{spec}{An object returned by either \code{\link{specify}} or \code{\link{hmm}}.}

\item{observation}{An optional logical indicating whether the observation model should be included in the description. It defaults to TRUE.}

\item{initial}{An optional logical indicating whether the initial distribution model should be included in the description. It defaults to TRUE.}

\item{transition}{An optional logical indicating whether the transition model should be included in the description. It defaults to TRUE.}

\item{print}{An optional logical indicating whether the description should be printing out.}
}
\value{
A character string with the model description.
}
\description{
The function creates a user-friendly text describing any of the three elements of the model. It includes the hidden states, variables, densities, bounds, priors, and fixed parameters. It also records environment details for easier reproducibility (package version, R version, time, OS).
}
\examples{
mySpec <- hmm(
  K = 2, R = 1,
  observation = Gaussian(
    mu    = Gaussian(0, 10),
    sigma = Student(
      mu = 0, sigma = 10, nu = 1, bounds = list(0, NULL)
    )
  ),
  initial     = Dirichlet(alpha = c(1, 1)),
  transition  = Dirichlet(alpha = c(1, 1)),
  name = "Univariate Gaussian Hidden Markov Model"
)

explain(mySpec)
}
