% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{explain.Specification}
\alias{explain.Specification}
\title{#'}
\usage{
\method{explain}{Specification}(spec, observation = TRUE,
  initial = TRUE, transition = TRUE, print = TRUE)
}
\description{
#'
}
