% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{explain_density}
\alias{explain_density}
\title{Explain a Density object in human readable format.}
\usage{
explain_density(x, print = FALSE)
}
\arguments{
\item{x}{A Density object.}

\item{print}{An optional logical indicating whether the description should be printing out.}
}
\value{
A character vector.
}
\description{
Explain a Density object in human readable format.
}
\examples{
explain_density(Gaussian(0, 1), print = TRUE)
}
