% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityTransitionFixed.R
\name{explain_density.TransitionFixed}
\alias{explain_density.TransitionFixed}
\title{Explain a Density object in human readable format.}
\usage{
\method{explain_density}{TransitionFixed}(x)
}
\arguments{
\item{x}{A \code{\link{Density}} object.}
}
\value{
A character string.
}
\description{
Explain a Density object in human readable format.
}
\keyword{internal}
