% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{explain_initial}
\alias{explain_initial}
\title{Create an outline of the initial model.}
\usage{
explain_initial(spec, ...)
}
\arguments{
\item{spec}{An object returned by the \code{\link{spec}}) function.}
}
\value{
A character vector with an outline of the initial model.
}
\description{
Create an outline of the initial model.
}
\seealso{
Other explain: \code{\link{explain_observation}},
  \code{\link{explain_transition}}
}
\concept{explain}
