% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{explain_observation}
\alias{explain_observation}
\title{Create an outline of the observation model.}
\usage{
explain_observation(spec)
}
\arguments{
\item{spec}{An object returned by the \code{\link{specify}}) function.}
}
\value{
A character vector with an outline of the observation model.
}
\description{
Create an outline of the observation model.
}
\seealso{
Other explain
#': \code{\link{explain_initial}},
  \code{\link{explain_transition}}
}
\concept{explain
#'}
\keyword{internal}
