% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{explain_transition}
\alias{explain_transition}
\title{Create an outline of the transition model.}
\usage{
explain_transition(spec)
}
\arguments{
\item{spec}{An object returned by the \code{\link{specify}}) function.}
}
\value{
A character vector with an outline of the transition model.
}
\description{
Create an outline of the transition model.
}
\seealso{
Other explain: \code{\link{explain_initial}},
  \code{\link{explain_observation}}
}
\concept{explain}
\keyword{internal}
