% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizationList.R
\name{extract_best}
\alias{extract_best}
\title{Return the optimization object for the run with the hightest posterior log density.}
\usage{
extract_best(stanoptimList)
}
\arguments{
\item{stanoptimList}{An object returned by \code{\link{optimizing}} when called with `nRuns` greater than one and `keep = "all"`.}
}
\value{
A named matrix with one row for each time the optimization procedure was run. Columns include the seed, the estimated log posterior evaluated at the maximum, the code returned by the optimization algorithm, the time elapsed (user, system and total) in seconds, and the estimated value of the parameters selected in
}
\description{
It doesn't do anything fancy but it's a neat quality-of-life feature :).
}
