% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{extract_gamma}
\alias{extract_gamma}
\title{Extract the estimates of the smoothed probability (gamma).}
\usage{
extract_gamma(fit, ...)
}
\arguments{
\item{fit}{An object returned by either \code{\link{sampling}} or \code{\link{optimizing}}.}
}
\value{
A numeric array with four dimensions [N, M, K, T]: number of iterations \emph{N}, number of chains \emph{M}, number of hidden states \emph{K}, length of the time series {T}. If the argument \emph{chain} was set, the chain dimension is dropped and the function returns a three-dimensional numeric array [N, K, T]. If the argument \emph{reduce} was set to a function returning a vector of size \emph{n}, the returned array has dimension [n, M, K, T]. When \emph{n} is simply one (e.g. \code{\link{median}}), the number-of-iterations dimension is dropped and the function returns a three-dimensional numeric array [M, K, T].
}
\description{
The extract_* family of methods returns elements from a fit object (including fixed and estimated quantities, as well as internal elements of the fit object).
}
\details{
The extract_family is the safest way to access the generated samples (filtered probability \emph{alpha}, smoothed probability \emph{gamma}, most likely hidden path \emph{zstar}, observation samples from the posterior predictive density \emph{ypred}, hidden path samples from the posterior predictive density \emph{zpred}, simulated observations \emph{ysim}), the fixed pre-specified quantities (observations \emph{y}, time series length \emph{T}, number of hidden states \emph{K}, dimension of the observation vector \emph{R}), and other internal elements stored in a fit object (dataset in list format \emph{data}, specification object \emph{spec}, path to the underlying Stan file \emph{filename}). Note that these functions work transparently with both Markov-chain Monte Carlo samples returned by \code{\link{sampling}} and maximum a posteriori estimates returned by \code{\link{optimizing}}.
}
\seealso{
Other extract: \code{\link{extract_K}},
  \code{\link{extract_R}}, \code{\link{extract_T}},
  \code{\link{extract_alpha}}, \code{\link{extract_data}},
  \code{\link{extract_filename}},
  \code{\link{extract_obs_parameters}},
  \code{\link{extract_parameters}},
  \code{\link{extract_spec}}, \code{\link{extract_ypred}},
  \code{\link{extract_ysim}}, \code{\link{extract_y}},
  \code{\link{extract_zpred}}, \code{\link{extract_zstar}},
  \code{\link{extract}}
}
\concept{extract}
