% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfit.R
\name{extract_n_chains}
\alias{extract_n_chains}
\title{Extract the number of chains (M).}
\usage{
extract_n_chains(stanfit)
}
\arguments{
\item{fit}{An object returned by either \code{\link{sampling}}.}
}
\value{
The number of chains used (M) to fit the model.
}
\description{
Extract the number of chains (M).
}
