% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfit.R
\name{extract_n_thin.stanfit}
\alias{extract_n_thin.stanfit}
\title{Extract the thinning periodicity.}
\usage{
\method{extract_n_thin}{stanfit}(stanfit)
}
\arguments{
\item{stanfit}{An object returned by either \code{\link{fit}} or \code{\link{draw_samples}}.}
}
\value{
The the thinning periodicity.
}
\description{
Extract the thinning periodicity.
}
\keyword{internal}
