% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfit.R
\name{extract_n_warmup.stanfit}
\alias{extract_n_warmup.stanfit}
\title{Extract the number of warmup iterations.}
\usage{
\method{extract_n_warmup}{stanfit}(stanfit)
}
\arguments{
\item{stanfit}{An object returned by either \code{\link{fit}} or \code{\link{draw_samples}}.}
}
\value{
The number of warmup iterations.
}
\description{
Extract the number of warmup iterations.
}
\keyword{internal}
