% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitIntegration.R
\name{extract_parameter_names}
\alias{extract_parameter_names}
\title{Extract the names of all the model parameters in the fit object.}
\usage{
extract_parameter_names(fit)
}
\arguments{
\item{fit}{An object returned by either \code{\link{sampling}} or \code{\link{optimizing}}.}
}
\value{
A character vector with the selected model parameter names.
}
\description{
Extract the names of all the model parameters in the fit object.
}
\seealso{
See \code{\link{select_parameters}} for extracting the names of part of the model.
}
