% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{extract_parameter_names.stanfit}
\alias{extract_parameter_names.stanfit}
\title{Extract the names of all the model parameters in the fit object.}
\usage{
\method{extract_parameter_names}{stanfit}(fit)
}
\arguments{
\item{fit}{An object returned by either \code{\link{draw_samples}} or \code{\link{optimizing}}.}
}
\value{
A character vector with the selected model parameter names.
}
\description{
Extract the names of all the model parameters in the fit object.
}
\seealso{
See \code{\link{select_parameters}} for extracting the names of part of the model.
}
\keyword{internal}
