% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_parameters}
\alias{extract_parameters}
\title{Extract the estimates of the model parameters (observation, transition, and initial models).}
\usage{
extract_parameters(fit, observation = TRUE, initial = TRUE,
  transition = TRUE, ...)
}
\arguments{
\item{fit}{An object returned by either \code{\link{drawSamples}} or \code{\link{optimizing}}.}

\item{observation}{A boolean indicating whether the observation model parameters should be returned.}

\item{initial}{A boolean indicating whether the initial model parameters should be returned.}

\item{transition}{A boolean indicating whether the transition model parameters should be returned.}

\item{...}{Arguments to be passed to rstan's \code{\link[rstan]{extract}} if the object \emph{fit} was returned by \code{\link{fit}} or \code{\link{drawSamples}}.}
}
\value{
A named list with as many elements as model parameters. Each element is a numeric array with dimensions [N, M, ...]: number of iterations \emph{N}, number of chains \emph{M}, and one or more dimensions related to the characteristics of the parameter. If the argument \emph{chain} was set, the chain dimension is dropped and the function returns a numeric array with dimensions [N, ...]. If the argument \emph{reduce} was set to a function returning a vector of size \emph{n}, the returned array has dimension [n, M, ...]. When \emph{n} is simply one (e.g. \code{\link{median}}), the number-of-iterations dimension is dropped and the function returns a three-dimensional numeric array [M, ...].
}
\description{
Extract the estimates of the model parameters (observation, transition, and initial models).
}
\seealso{
Other extract
#': \code{\link{extract_K}},
  \code{\link{extract_R}}, \code{\link{extract_T}},
  \code{\link{extract_alpha}}, \code{\link{extract_data}},
  \code{\link{extract_filename}},
  \code{\link{extract_gamma}}, \code{\link{extract_spec}},
  \code{\link{extract_ypred}}, \code{\link{extract_ysim}},
  \code{\link{extract_y}}, \code{\link{extract_zpred}},
  \code{\link{extract_zsim}}, \code{\link{extract_zstar}}
}
\concept{extract
#'}
