% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfit.R
\docType{methods}
\name{extract_sample_size}
\alias{extract_sample_size}
\alias{extract_sample_size,stanfit-method}
\title{Extract the number of iterations kept after warmup.}
\usage{
extract_sample_size(stanfit)

\S4method{extract_sample_size}{stanfit}(stanfit)
}
\arguments{
\item{stanfit}{An object returned by either \code{\link{fit}} or \code{\link{draw_samples}}.}
}
\value{
The number of iterations kept after warmup.
}
\description{
Extract the number of iterations kept after warmup.
}
