% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{extract_seed.Optimization}
\alias{extract_seed.Optimization}
\title{Extract the seed used to fit the model.}
\usage{
\method{extract_seed}{Optimization}(fit)
}
\arguments{
\item{fit}{An object returned by either \code{\link{draw_samples}} or \code{\link{optimizing}}.}
}
\value{
An integer with the seed used to fit the model.
}
\description{
Extract the seed used to fit the model.
}
\keyword{internal}
