% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{extract_stanmodel.stanmodel}
\alias{extract_stanmodel.stanmodel}
\title{Load the underlying Stan code into an IDE or browser.}
\usage{
extract_stanmodel.stanmodel(fit)
}
\value{
No return value.
}
\description{
The behavior is platform dependent.
}
\seealso{
\code{\link{browseURL}}.
}
\keyword{internal}
