% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_T}
\alias{extract_T}
\title{Extract the length of the time series \emph{T}.}
\usage{
extract_T(fit)
}
\arguments{
\item{fit}{An object returned by either \code{\link{draw_samples}} or \code{\link{optimizing}}.}
}
\value{
An integer with the length of the time series \emph{T}.
}
\description{
The extract_* family of methods returns elements from a fit object (including fixed and estimated quantities, as well as internal elements of the fit object).
}
\details{
The extract_* family is the safest way to access the generated samples (filtered probability \emph{alpha}, smoothed probability \emph{gamma}, most likely hidden path \emph{zstar}, observation samples from the posterior predictive density \emph{ypred}, hidden path samples from the posterior predictive density \emph{zpred}, simulated observations \emph{ysim}), the fixed pre-specified quantities (observations \emph{y}, time series length \emph{T}, number of hidden states \emph{K}, dimension of the observation vector \emph{R}), Markov-chain Monte Carlo specific information (convergence and posterior predictive diagnostics \emph{diagnostics}, number of chains \emph{n_chains}, total iterations \emph{n_iterations}, warm-up iterations \emph{n_warmup}, thinning \emph{n_thin}, sample size \emph{sample_size}), and other internal elements stored in a fit object (\code{\link{Specification}} object \emph{spec}, dataset in list format \emph{data}, date \emph{date}, time elapsed for fitting \emph{time}, seed \emph{seed}, path to the underlying Stan file \emph{filename}). Except for MCMC-specific functions, these all work interchangeably with both Markov-chain Monte Carlo samples returned by \code{\link{draw_samples}} and maximum a posteriori estimates returned by \code{\link{optimizing}}.
}
\note{
This function does not allow the arguments \emph{combine}, \emph{reduce}, and \emph{chain}.
}
\seealso{
Other extract: \code{\link{extract_K}},
  \code{\link{extract_R}}, \code{\link{extract_alpha}},
  \code{\link{extract_data}},
  \code{\link{extract_filename}},
  \code{\link{extract_gamma}}, \code{\link{extract_model}},
  \code{\link{extract_obs_parameters}},
  \code{\link{extract_parameters}},
  \code{\link{extract_spec}}, \code{\link{extract_ypred}},
  \code{\link{extract_ysim}}, \code{\link{extract_y}},
  \code{\link{extract_zpred}}, \code{\link{extract_zsim}},
  \code{\link{extract_zstar}}, \code{\link{extract}}
}
\concept{extract}
