% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitIntegration.R
\name{extract_time}
\alias{extract_time}
\title{Extract the time elapsed when fitting the model.}
\usage{
extract_time(fit)
}
\arguments{
\item{fit}{An object returned by either \code{\link{sampling}} or \code{\link{optimizing}}.}
}
\value{
For objects fit via \code{\link{sampling}}, a named numeric matrix with two columns (warm-up and sample time) and one row per chain. For objects fit via \code{\link{optimizing}}, a ...?
}
\description{
Extract the time elapsed when fitting the model.
}
