% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{extract_ysim}
\alias{extract_ysim}
\title{Extract the simulated sample of the observation variable (ysim).}
\usage{
extract_ysim(fit, n = NULL, ...)
}
\arguments{
\item{fit}{An object returned by either \code{\link{sampling}} or \code{\link{optimizing}}.}
}
\value{
A numeric array with four dimensions [N, M, T, R]: number of iterations \emph{N}, number of chains \emph{M}, length of the time series {T}, the dimension of the observation vector \emph{R}. If the argument \emph{chain} was set, the chain dimension is dropped and the function returns a three-dimensional numeric array [N, T, R]. If the argument \emph{reduce} was set to a function returning a vector of size \emph{n}, the returned array has dimension [n, M, T, R]. When \emph{n} is simply one (e.g. \code{\link{median}}), the number-of-iterations dimension is dropped and the function returns a three-dimensional numeric array [M, T, R].
}
\description{
If the \emph{fit} object was created via the \code{\link{sim}} method, this function returns a sample of the observation variable from the prior predictive densitiy (i.e. using random values of the parameters). If a dataset was provided when fitting the model instead, the sample is simulated from the posterior predictive density. In the latter case, calling either \code{\link{extract_ysim}} and \code{\link{extract_ypred}} are equivalent.
}
\seealso{
Other extract: \code{\link{extract_K}},
  \code{\link{extract_R}}, \code{\link{extract_T}},
  \code{\link{extract_alpha}}, \code{\link{extract_data}},
  \code{\link{extract_filename}},
  \code{\link{extract_gamma}},
  \code{\link{extract_obs_parameters}},
  \code{\link{extract_parameters}},
  \code{\link{extract_spec}}, \code{\link{extract_ypred}},
  \code{\link{extract_y}}, \code{\link{extract_zpred}},
  \code{\link{extract_zstar}}, \code{\link{extract}}
}
\concept{extract}
