% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{extract_zpred}
\alias{extract_zpred}
\title{Extract the sample of the hidden state path drawn from the posterior predictive density (zpred).}
\usage{
extract_zpred(fit, ...)
}
\arguments{
\item{fit}{An object returned by either \code{\link{sampling}} or \code{\link{optimizing}}.}
}
\value{
A numeric array with three dimensions [N, M, T]: number of iterations \emph{N}, number of chains \emph{M}, length of the time series {T}, the dimension of the observation vector \emph{R}. If the argument \emph{chain} was set, the chain dimension is dropped and the function returns a two-dimensional numeric array [N, T]. If the argument \emph{reduce} was set to a function returning a vector of size \emph{n}, the returned array has dimension [n, M, T]. When \emph{n} is simply one (e.g. \code{\link{median}}), the number-of-iterations dimension is dropped and the function returns a two-dimensional numeric array [M, T].
}
\description{
Extract the sample of the hidden state path drawn from the posterior predictive density (zpred).
}
\seealso{
Other extract: \code{\link{extract_K}},
  \code{\link{extract_R}}, \code{\link{extract_T}},
  \code{\link{extract_alpha}}, \code{\link{extract_data}},
  \code{\link{extract_filename}},
  \code{\link{extract_gamma}},
  \code{\link{extract_obs_parameters}},
  \code{\link{extract_parameters}},
  \code{\link{extract_spec}}, \code{\link{extract_ypred}},
  \code{\link{extract_ysim}}, \code{\link{extract_y}},
  \code{\link{extract_zstar}}, \code{\link{extract}}
}
\concept{extract}
