% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{fit}
\alias{fit}
\title{Fit a model by MCMC}
\usage{
fit(spec, y, x = NULL, u = NULL, v = NULL, ...)
}
\arguments{
\item{spec}{An object returned by either \code{\link{specify}} or \code{\link{hmm}}.}

\item{...}{Arguments to be passed to rstan's \code{\link[rstan]{stan}}.}
}
\value{
An object of S4 class stanfit with some additional attributes (the dataset \emph{data}, the name of the Stan code file \emph{filename}, and the specification object \emph{spec}). This object is completely compatible with all other functions.
}
\description{
Fit a model by MCMC
}
\seealso{
See rstan's \code{\link[rstan]{stan}} and \code{\link[rstan]{sampling}} for further details on tunning the MCMC algorithm.
}
