% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{fixedParameters}
\alias{fixedParameters}
\title{Write the Stan code lines related with Density fixed parameters.}
\usage{
fixedParameters(x)
}
\arguments{
\item{x}{A \code{\link{Density}} object.}
}
\value{
A character string.
}
\description{
Write the Stan code lines related with Density fixed parameters.
}
\keyword{internal}
