% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityRegGaussian.R
\name{fixedParameters.RegGaussian}
\alias{fixedParameters.RegGaussian}
\title{Write the Stan code lines related with Density fixed parameters.}
\usage{
\method{fixedParameters}{RegGaussian}(x)
}
\arguments{
\item{x}{A \code{\link{Density}} object.}
}
\value{
A character vector.
}
\description{
Write the Stan code lines related with Density fixed parameters.
}
