% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityTransitionFixed.R
\name{fixedParameters.TransitionFixed}
\alias{fixedParameters.TransitionFixed}
\title{Write the Stan code lines related with Density fixed parameters.}
\usage{
\method{fixedParameters}{TransitionFixed}(x)
}
\arguments{
\item{x}{A \code{\link{Density}} object.}
}
\value{
A character string.
}
\description{
Write the Stan code lines related with Density fixed parameters.
}
\keyword{internal}
