% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityInitialFixed.R
\name{freeParameters.InitialFixed}
\alias{freeParameters.InitialFixed}
\title{Write the Stan code lines related with Density free parameters.}
\usage{
\method{freeParameters}{InitialFixed}(x)
}
\arguments{
\item{x}{A \code{\link{Density}} object.}
}
\value{
A character string.
}
\description{
Write the Stan code lines related with Density free parameters.
}
\keyword{internal}
