% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityInitialSoftmax.R
\name{freeParameters.InitialSoftmax}
\alias{freeParameters.InitialSoftmax}
\title{Write the Stan code lines related with Density free parameters.}
\usage{
\method{freeParameters}{InitialSoftmax}(x)
}
\arguments{
\item{x}{A \code{\link{Density}} object.}
}
\value{
A character vector.
}
\description{
Write the Stan code lines related with Density free parameters.
}
