% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityGamma.R
\name{generated.GammaDensity}
\alias{generated.GammaDensity}
\title{Write the Stan code lines related with Density generated quantities.}
\usage{
\method{generated}{GammaDensity}(x)
}
\arguments{
\item{x}{A \code{\link{Density}} object.}
}
\value{
A character string.
}
\description{
Write the Stan code lines related with Density generated quantities.
}
\keyword{internal}
