% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityRegBernoulliLogit.R
\name{generated.RegBernoulliLogit}
\alias{generated.RegBernoulliLogit}
\title{Write the Stan code lines related with Density generated quantities.}
\usage{
\method{generated}{RegBernoulliLogit}(x)
}
\arguments{
\item{x}{A \code{\link{Density}} object.}
}
\value{
A character vector.
}
\description{
Write the Stan code lines related with Density generated quantities.
}
