% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{get_default_theme}
\alias{get_default_theme}
\title{Return the default theme.}
\usage{
get_default_theme()
}
\value{
A named list with the theme.
}
\description{
Return the default theme.
}
\seealso{
Other visualization functions: \code{\link{get_current_theme}},
  \code{\link{modify_theme_entry}},
  \code{\link{plot_ppredictive}},
  \code{\link{plot_series}},
  \code{\link{plot_state_probability}}, \code{\link{theme}}
}
\concept{visualization functions}
