% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{get_plot_theme}
\alias{get_plot_theme}
\title{Return the current theme for visualizations.}
\usage{
get_plot_theme()
}
\value{
A named list with the theme for visualizations.
}
\description{
Return the current theme for visualizations.
}
\keyword{internal}
