% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{get_print_settings}
\alias{get_print_settings}
\title{Return the current theme for text printouts.}
\usage{
get_print_settings()
}
\value{
A named list with the theme for text printouts..
}
\description{
Return the current theme for text printouts.
}
\keyword{internal}
