% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{getFreeParameters}
\alias{getFreeParameters}
\title{Return the names of the free parameters in the Density object.}
\usage{
getFreeParameters(x)
}
\arguments{
\item{x}{A \code{\link{Density}} object.}
}
\value{
A vector of character strings.
}
\description{
Return the names of the free parameters in the Density object.
}
\examples{
\dontrun{
getFreeParameters(
  Gaussian(
    mu    = Gaussian(mu = 0, sigma = 10),
    sigma = 1
  )
)
}
}
\keyword{internal}
