% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityMVGaussian.R
\name{getParameterNames.MVGaussian}
\alias{getParameterNames.MVGaussian}
\title{Return the names of both free and fixed parameters in the Density object.}
\usage{
\method{getParameterNames}{MVGaussian}(x)
}
\arguments{
\item{x}{#' @param x A \code{\link{Density}} object.}
}
\value{
A vector of character vectors.
}
\description{
Return the names of both free and fixed parameters in the Density object.
}
\examples{
\dontrun{
getParameterNames(
  Gaussian(
    mu    = Gaussian(mu = 0, sigma = 10),
    sigma = 1
  )
)
}
}
