% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityMVStudent.R
\name{getParameterNames.MVStudent}
\alias{getParameterNames.MVStudent}
\title{Return the names of both free and fixed parameters in the Density object.}
\usage{
\method{getParameterNames}{MVStudent}(x)
}
\arguments{
\item{x}{A \code{\link{Density}} object.}
}
\value{
A vector of character strings.
}
\description{
Return the names of both free and fixed parameters in the Density object.
}
\examples{
\dontrun{
getParameterNames(
  Gaussian(
    mu    = Gaussian(mu = 0, sigma = 10),
    sigma = 1
  )
)
}
}
\keyword{internal}
