% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{getParameterNames}
\alias{getParameterNames}
\title{Return the names of both free and fixed parameters in the Density object.}
\usage{
getParameterNames(x)
}
\arguments{
\item{x}{A Density object.}
}
\value{
A vector of character vectors.
}
\description{
Return the names of both free and fixed parameters in the Density object.
}
\examples{
getParameterNames(
  Gaussian(
    mu    = Gaussian(mu = 0, sigma = 10),
    sigma = 1
  )
)
}
