% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{is.DensityList}
\alias{is.DensityList}
\title{Check if it is a \code{\link{DensityList}} object.}
\usage{
is.DensityList(x)
}
\arguments{
\item{x}{An object.}
}
\value{
TRUE if the object is a DensityList object, FALSE otherwise.
}
\description{
Check if it is a \code{\link{DensityList}} object.
}
