% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{is.discrete}
\alias{is.discrete}
\title{Check if it is a Density object for discrete random variables.}
\usage{
is.discrete(x)
}
\arguments{
\item{x}{a Density object.}
}
\value{
TRUE if the object is meant to represent a discrete random variables, FALSE otherwise.
}
\description{
Check if it is a Density object for discrete random variables.
}
\examples{
is.discrete(
  Gaussian(0, 1)
)

is.discrete(
  Poisson(lambda = GammaDensity(1, 1))
)
}
