% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificationInternals.R
\name{is.discrete.Specification}
\alias{is.discrete.Specification}
\title{Check if it is a Density object for discrete random variables.}
\usage{
\method{is.discrete}{Specification}(x)
}
\arguments{
\item{x}{#' @param x A \code{\link{Density}} object.}
}
\value{
TRUE if the object is meant to represent a discrete random variables, FALSE otherwise.
}
\description{
Check if it is a Density object for discrete random variables.
}
