% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{is.equal}
\alias{is.equal}
\title{Check if it is a Density object for a parameter with equal value across hidden states.}
\usage{
is.equal(x)
}
\arguments{
\item{x}{A \code{\link{Density}} object.}
}
\value{
TRUE if the object is meant to represent a parameter shared across states, FALSE otherwise.
}
\description{
Check if it is a Density object for a parameter with equal value across hidden states.
}
\keyword{internal}
