% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificationInternals.R
\name{is.FixedInitial}
\alias{is.FixedInitial}
\title{Check if it is a fixed initial probability object.}
\usage{
is.FixedInitial(x)
}
\arguments{
\item{x}{A \code{\link{Density}} or a \code{\link{Specification}} object.}
}
\value{
TRUE if the object is a fixed initial probability object, FALSE otherwise.
}
\description{
Check if it is a fixed initial probability object.
}
\keyword{internal}
