% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificationInternals.R
\name{is.FixedTransition.Specification}
\alias{is.FixedTransition.Specification}
\title{Check if it is a fixed transition probability object.}
\usage{
\method{is.FixedTransition}{Specification}(x)
}
\arguments{
\item{x}{A \code{\link{Density}} or a \code{\link{Specification}} object.}
}
\value{
TRUE if the object is a fixed transition probability object, FALSE otherwise.
}
\description{
Check if it is a fixed transition probability object.
}
\keyword{internal}
