% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{is.link.Density}
\alias{is.link.Density}
\title{Check if it is a Density object for the transition model.}
\usage{
\method{is.link}{Density}(x)
}
\arguments{
\item{x}{A \code{\link{Density}} object.}
}
\value{
TRUE if the object is meant for transition models, FALSE otherwise.
}
\description{
Check if it is a Density object for the transition model.
}
\keyword{internal}
