% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{is.multivariate.Density}
\alias{is.multivariate.Density}
\title{Check if it is a Density object for multivariate random variables.}
\usage{
\method{is.multivariate}{Density}(x)
}
\arguments{
\item{x}{#' @param x A \code{\link{Density}} object.}
}
\value{
TRUE if the object is meant to represent a multivariate random variables, FALSE otherwise.
}
\description{
Check if it is a Density object for multivariate random variables.
}
