% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{is.multivariate.DensityList}
\alias{is.multivariate.DensityList}
\title{Check if it is a Density object for multivariate random variables.}
\usage{
\method{is.multivariate}{DensityList}(x)
}
\arguments{
\item{x}{A \code{\link{Density}} object.}
}
\value{
TRUE if the object is meant to represent a multivariate random variables, FALSE otherwise.
}
\description{
Check if it is a Density object for multivariate random variables.
}
\keyword{internal}
