% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{is.ordered.Density}
\alias{is.ordered.Density}
\title{Check if it is a Density object for an ordered set of parameters.}
\usage{
\method{is.ordered}{Density}(x)
}
\arguments{
\item{x}{A \code{\link{Density}} object.}
}
\value{
TRUE if the object is meant to represent an ordered set of parameters, FALSE otherwise.
}
\description{
Check if it is a Density object for an ordered set of parameters.
}
\keyword{internal}
