% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{is.Specification}
\alias{is.Specification}
\title{Check if it is a \code{\link{Specification}} object.}
\usage{
is.Specification(x)
}
\arguments{
\item{x}{An object.}
}
\value{
TRUE if the object is a Specification object, FALSE otherwise.
}
\description{
Check if it is a \code{\link{Specification}} object.
}
\keyword{internal}
