% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfit.R
\name{is.stanfit}
\alias{is.stanfit}
\title{Check if it is an object created by \code{\link{drawSamples}}.}
\usage{
is.stanfit(x)
}
\arguments{
\item{x}{An object.}
}
\value{
TRUE if it is an object created by \code{\link{drawSamples}}.
}
\description{
Check if it is an object created by \code{\link{drawSamples}}.
}
\keyword{internal}
