% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.R
\name{is.stanoptim}
\alias{is.stanoptim}
\title{Verify that the object was created by \code{\link{optimizing}}.}
\usage{
is.stanoptim(x)
}
\arguments{
\item{x}{An object.}
}
\value{
TRUE if it is an object created by \code{\link{optimizing}}.
}
\description{
Verify that the object was created by \code{\link{optimizing}}.
}
\keyword{internal}
