% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityInitialSoftmax.R
\name{is.TVInitial.InitialSoftmax}
\alias{is.TVInitial.InitialSoftmax}
\title{Check if it is a time-varying initial distribution object.}
\usage{
\method{is.TVInitial}{InitialSoftmax}(x)
}
\arguments{
\item{x}{A \code{\link{Density}} or a \code{\link{Specification}} object.}
}
\value{
TRUE if the object is a time-varying initial distribution object, FALSE otherwise.
}
\description{
Check if it is a time-varying initial distribution object.
}
\keyword{internal}
