% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityInitialSoftmax.R
\name{link.InitialSoftmax}
\alias{link.InitialSoftmax}
\title{Write the Stan code lines related with Density link function.}
\usage{
\method{link}{InitialSoftmax}(x)
}
\arguments{
\item{x}{A \code{\link{Density}} object.}
}
\value{
A character string.
}
\description{
Write the Stan code lines related with Density link function.
}
\keyword{internal}
