% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{link}
\alias{link}
\title{Write the Stan code lines related with Density link function.}
\usage{
link(x)
}
\arguments{
\item{x}{A \code{\link{Density}} object.}
}
\value{
A character string.
}
\description{
Write the Stan code lines related with Density link function.
}
\keyword{internal}
