% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{LinkDensity}
\alias{LinkDensity}
\title{Create a representation of a link function that may be used only to specify
a transition model.}
\usage{
LinkDensity(name, ...)
}
\arguments{
\item{name}{A character vector with the name of the density.}

\item{...}{Other arguments for the density.

#'}
}
\value{
A Density object.
}
\description{
Create a representation of a link function that may be used only to specify
a transition model.
}
\concept{LinkDensity}
