% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityBernoulli.R
\name{logLike.Bernoulli}
\alias{logLike.Bernoulli}
\title{Write the Stan code lines related with Density log likelihood.}
\usage{
\method{logLike}{Bernoulli}(x)
}
\arguments{
\item{x}{A \code{\link{Density}} object.}
}
\value{
A character string.
}
\description{
Write the Stan code lines related with Density log likelihood.
}
\keyword{internal}
