% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityMVGaussian.R
\name{logLike.MVGaussian}
\alias{logLike.MVGaussian}
\title{Write the Stan code lines related with Density log likelihood.}
\usage{
\method{logLike}{MVGaussian}(x)
}
\arguments{
\item{x}{A \code{\link{Density}} object.}
}
\value{
A character vector.
}
\description{
Write the Stan code lines related with Density log likelihood.
}
