% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_text_header}
\alias{make_text_header}
\title{Make a character string with a header.}
\usage{
make_text_header(text)
}
\value{
A character string with a header.
}
\description{
This function formats the text into a header.
}
\keyword{internal}
