% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_text_subheader}
\alias{make_text_subheader}
\title{Make a character string with a subheader.}
\usage{
make_text_subheader(text)
}
\value{
A character string with a subheader.
}
\description{
This function formats the text into a subheader.
}
\keyword{internal}
