% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{match_parameter_names}
\alias{match_parameter_names}
\title{Return the name of the parameters matching the pattern.}
\usage{
match_parameter_names(fit, pars)
}
\arguments{
\item{fit}{An object returned by either \code{\link{draw_samples}} or \code{\link{optimizing}}.}

\item{pars}{Regular expressions against which parameter names should be matched.}
}
\value{
A character vector with the names of matching model parameters.
}
\description{
Return the name of the parameters matching the pattern.
}
\keyword{internal}
