% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{modify_theme_entry}
\alias{modify_theme_entry}
\title{Modifies an entry in the current theme.}
\usage{
modify_theme_entry(section, key, value)
}
\arguments{
\item{section}{A character string with the name of the section (either theme or print).}

\item{key}{A character string with the key.}

\item{value}{A character string with the value.}
}
\value{
A named list with the theme.
}
\description{
Modifies an entry in the current theme.
}
\seealso{
Other visualization functions: \code{\link{get_current_theme}},
  \code{\link{get_default_theme}},
  \code{\link{plot_ppredictive}},
  \code{\link{plot_series}},
  \code{\link{plot_state_probability}}, \code{\link{theme}}
}
\concept{visualization functions}
