% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{MultivariateDensity}
\alias{MultivariateDensity}
\title{Create a representation of a probability mass or density function for a
continuous multivariate random variable. It can be used to specify either a
prior distribution for a model parameter or a likelihood function for an
observation model.}
\usage{
MultivariateDensity(name, ...)
}
\arguments{
\item{name}{A character vector with the name of the density.}

\item{...}{Other arguments for the density.

#'}
}
\value{
A Density object.
}
\description{
Create a representation of a probability mass or density function for a
continuous multivariate random variable. It can be used to specify either a
prior distribution for a model parameter or a likelihood function for an
observation model.
}
\concept{MultivariateDensity}
