% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityMVGaussian.R
\name{MVGaussian}
\alias{MVGaussian}
\title{Multivariate Gaussian density (Multivariate, continuous, unbounded space)}
\usage{
MVGaussian(mu = NULL, sigma = NULL, bounds = list(NULL, NULL),
  trunc = list(NULL, NULL), k = NULL, r = NULL, param = NULL)
}
\arguments{
\item{mu}{Either a fixed value or a prior density for the mean vector.}

\item{sigma}{Either a fixed value or a prior density for the covariance matrix.}

\item{bounds}{(optional) A list with two elements specifying the lower and upper bound for the parameter space. Use either a fixed value for a finite bound or NULL for no bounds. It defaults to an unbounded parameter space.}

\item{trunc}{(optional) A list with two elements specifying the lower and upper bound for the domain of the density function. Use either a fixed value for a finite bound or NULL for no truncation. It defaults to an unbounded domain.}

\item{k}{(optional) The number of the hidden state for which this density should be used. This argument is mostly for internal use: you should not use it unless you are acquainted with the internals of this software.}

\item{r}{(optional) The dimension of the observation vector dimension for which this density should be used. This argument is mostly for internal use: you should not use it unless you are acquainted with the internals of this software.}

\item{param}{(optional) The name of the parameter. This argument is mostly for internal use: you should not use it unless you are acquainted with the internals of this software.}
}
\value{
A Density object.
}
\description{
Multivariate Gaussian density (Multivariate, continuous, unbounded space)
}
\examples{
# With fixed values for the parameters
MVGaussian(
  mu    = c(0, 0),
  sigma = matrix(c(1, 0, 0, 1), 2, 2)
)

# With priors for the parameters
MVGaussian(
  mu    = MVGaussian(mu = c(0, 0), sigma = matrix(c(1, 0, 0, 1), 2, 2)),
  sigma = InverseWishart(nu = 5, sigma = matrix(c(1, 0, 0, 1), 2, 2))
)
}
\seealso{
Other Density: \code{\link{Bernoulli}}, \code{\link{Beta}},
  \code{\link{Binomial}}, \code{\link{Categorical}},
  \code{\link{Cauchy}}, \code{\link{CholeskyLKJCor}},
  \code{\link{Default}}, \code{\link{Density}},
  \code{\link{Dirichlet}}, \code{\link{Exponential}},
  \code{\link{GammaDensity}}, \code{\link{Gaussian}},
  \code{\link{InverseWishart}},
  \code{\link{MVGaussianCholeskyCor}},
  \code{\link{MVStudent}}, \code{\link{Multinomial}},
  \code{\link{NegativeBinomialLocation}},
  \code{\link{NegativeBinomial}}, \code{\link{Poisson}},
  \code{\link{Student}}, \code{\link{Wishart}}
}
\concept{Density}
