% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityMVGaussianCholeskyCor.R
\name{MVGaussianCholeskyCor}
\alias{MVGaussianCholeskyCor}
\title{Multivariate Gaussian density with Cholesky decomposition of the correlation matrix (Multivariate, continuous, unbounded space)}
\usage{
MVGaussianCholeskyCor(mu = NULL, L = NULL, bounds = list(NULL, NULL),
  trunc = list(NULL, NULL), k = NULL, r = NULL, param = NULL)
}
\arguments{
\item{mu}{Either a fixed value or a prior density for the mean vector.}

\item{L}{Either a fixed value or a prior density for the Cholesky factor of the correlation matrix.}

\item{bounds}{(optional) A list with two elements specifying the lower and upper bound for the parameter space. Use either a fixed value for a finite bound or NULL for no bounds. It defaults to an unbounded parameter space.}

\item{trunc}{(optional) A list with two elements specifying the lower and upper bound for the domain of the density function. Use either a fixed value for a finite bound or NULL for no truncation. It defaults to an unbounded domain.}

\item{k}{(optional) The number of the hidden state for which this density should be used. This argument is mostly for internal use: you should not use it unless you are acquainted with the internals of this software.}

\item{r}{(optional) The dimension of the observation vector dimension for which this density should be used. This argument is mostly for internal use: you should not use it unless you are acquainted with the internals of this software.}

\item{param}{(optional) The name of the parameter. This argument is mostly for internal use: you should not use it unless you are acquainted with the internals of this software.}
}
\value{
#' @param x A \code{\link{Density}} object.
}
\description{
Multivariate Gaussian density with Cholesky decomposition of the correlation matrix (Multivariate, continuous, unbounded space)
}
\examples{
# With priors for the parameters
MVGaussianCholeskyCor(
  mu    = MVGaussian(mu = c(0, 0), sigma = matrix(c(1, 0, 0, 1), 2, 2)),
  L     = CholeskyLKJCor(1)
)
}
\seealso{
Other Density
#': \code{\link{Bernoulli}},
  \code{\link{Beta}}, \code{\link{Cauchy}},
  \code{\link{CholeskyLKJCor}}, \code{\link{Default}},
  \code{\link{Dirichlet}}, \code{\link{Exponential}},
  \code{\link{GammaDensity}}, \code{\link{Gaussian}},
  \code{\link{InverseWishart}}, \code{\link{MVGaussian}},
  \code{\link{MVStudent}}, \code{\link{Multinomial}},
  \code{\link{NegativeBinomialLocation}},
  \code{\link{NegativeBinomial}}, \code{\link{Poisson}},
  \code{\link{Student}}, \code{\link{Wishart}}
}
\concept{Density
#'}
