% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{noLogLike}
\alias{noLogLike}
\title{Write the Stan code lines related with Density no log likelihood.}
\usage{
noLogLike(x)
}
\arguments{
\item{x}{A \code{\link{Density}} object.}
}
\value{
A character string.
}
\description{
Write the Stan code lines related with Density no log likelihood.
}
\keyword{internal}
