% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{optimizing.Specification}
\alias{optimizing.Specification}
\title{#'}
\usage{
\method{optimizing}{Specification}(spec, stanModel = NULL, y, x = NULL,
  u = NULL, v = NULL, nRuns = 1, keep = "best", nCores = 1,
  writeDir = tempdir(), ...)
}
\description{
#'
}
