% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{optimizing_all}
\alias{optimizing_all}
\title{Run several instances of the optimization algorithm.}
\usage{
optimizing_all(stanDots, nRuns, nCores)
}
\arguments{
\item{stanDots}{The arguments to the passed to rstan's \code{\link[rstan]{optimizing}}.}

\item{nRuns}{An optional integer with the number of initializations.}

\item{nCores}{An optional integer with the number of cores to be used. If equal to one, the instances are run sequentially. Otherwise, doParallel's backend is used for parallel computing. It defaults to one.}
}
\value{
An \emph{OptimizationList} object.
}
\description{
Note that this function returns the results of all the instances while \code{\link{optimizing_best}} only returns the converging instance with highest log posterior density.
}
\keyword{internal}
