% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{optimizing_run}
\alias{optimizing_run}
\title{Run one instance of the}
\usage{
optimizing_run(stanDots, n)
}
\arguments{
\item{stanDots}{The arguments to the passed to rstan's \code{\link[rstan]{optimizing}}.}

\item{n}{An integer with the number of the instance (i.e. the n-th time the algorithm is run on this model).}
}
\value{
An \code{\link{Optimization}} object.
}
\description{
Run one instance of the
}
\keyword{internal}
