% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{parameters}
\alias{parameters}
\title{Write the Stan code lines related with Density parameters.}
\usage{
parameters(x)
}
\arguments{
\item{x}{A \code{\link{Density}} object.}
}
\value{
A character vector.
}
\description{
Write the Stan code lines related with Density parameters.
}
