% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{+.Density}
\alias{+.Density}
\alias{+}
\title{Appends two Density objects.}
\usage{
x + y
}
\arguments{
\item{x}{A Density object (e.g. \code{\link{Gaussian}})}

\item{y}{A Density object (e.g. \code{\link{Gaussian}})}
}
\value{
A DensityList object.
}
\description{
Appends two Density objects.
}
\examples{
Gaussian() + Gaussian()
}
