% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{posterior_intervals}
\alias{posterior_intervals}
\title{Return a function that computes the posterior intervals.}
\usage{
posterior_intervals(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link{quantile}}.}
}
\value{
A scalar with same type as `x`.
}
\description{
Return a function that computes the posterior intervals.
}
\examples{
f <- posterior_intervals(probs = c(0.05, 0.95))
print(f(1:100))
}
