% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{posterior_mode}
\alias{posterior_mode}
\title{Return the posterior mode of a vector.}
\usage{
posterior_mode(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A scalar with same type as `x`.
}
\description{
Return the posterior mode of a vector.
}
\examples{
posterior_mode(sample(1:3, 20, replace = TRUE))
}
