% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.Density}
\alias{print.Density}
\title{Print a description of a Density object in a human friendly format.}
\usage{
\method{print}{Density}(x, print = TRUE, ...)
}
\arguments{
\item{x}{A \code{\link{Density}} object.}

\item{print}{An optional logical indicating whether the description should be printed out.}

\item{...}{Further arguments passed to \code{\link{cat}}.}
}
\value{
A character string with the density description.
}
\description{
Print a description of a Density object in a human friendly format.
}
