% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printOptimizing.R
\name{print.Optimization}
\alias{print.Optimization}
\title{Prints the result of the model in a human friendly format.}
\usage{
\method{print}{Optimization}(x, observation = TRUE, initial = TRUE,
  transition = TRUE, fixed = TRUE, diagnostics = TRUE,
  print = TRUE, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{optimizing}}.}

\item{observation}{An optional logical indicating whether the observation model should be included in the description. It defaults to TRUE.}

\item{initial}{An optional logical indicating whether the initial distribution model should be included in the description. It defaults to TRUE.}

\item{transition}{An optional logical indicating whether the transition model should be included in the description. It defaults to TRUE.}

\item{fixed}{An optional logical indicating whether the fixed parameters should be included in the description. It defaults to TRUE.}

\item{diagnostics}{An optional logical indicating whether convergence diagnostics should be included in the description. It defaults to TRUE.}

\item{print}{An optional logical indicating whether the description should be printing out.}

\item{...}{Further arguments passed to \code{\link{cat}}.}
}
\value{
A character string with the model description.
}
\description{
Prints the result of the model in a human friendly format.
}
