% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.stanfit}
\alias{print.stanfit}
\title{Prints the result of the model in a human friendly format.}
\usage{
\method{print}{stanfit}(x, posteriorInterval = c(0.025, 0.25, 0.5, 0.75,
  0.975), observation = TRUE, initial = TRUE, transition = TRUE,
  fixed = TRUE, print = TRUE, ...)
}
\arguments{
\item{x}{An object returned by either \code{\link{fit}} or \code{\link{draw_samples}}.}

\item{posteriorInterval}{An optional numeric vector with the quantilies of the posterior marginal distributions.}

\item{observation}{An optional logical indicating whether the observation model should be included in the description. It defaults to TRUE.}

\item{initial}{An optional logical indicating whether the initial distribution model should be included in the description. It defaults to TRUE.}

\item{transition}{An optional logical indicating whether the transition model should be included in the description. It defaults to TRUE.}

\item{fixed}{An optional logical indicating whether the fixed parameters should be included in the description. It defaults to TRUE.}

\item{print}{An optional logical indicating whether the description should be printing out.}

\item{...}{Further arguments passed to \code{\link{cat}}.}
}
\value{
A character string with the model description.
}
\description{
Prints the result of the model in a human friendly format.
}
