% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityBernoulli.R
\name{prior.Bernoulli}
\alias{prior.Bernoulli}
\title{Write the Stan code lines related with Density priors.}
\usage{
\method{prior}{Bernoulli}(x)
}
\arguments{
\item{x}{A \code{\link{Density}} object.}
}
\value{
A character string.
}
\description{
Write the Stan code lines related with Density priors.
}
\keyword{internal}
