% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityNegativeBinomial.R
\name{prior.NegativeBinomial}
\alias{prior.NegativeBinomial}
\title{Write the Stan code lines related with Density priors.}
\usage{
\method{prior}{NegativeBinomial}(x)
}
\arguments{
\item{x}{A \code{\link{Density}} object.}
}
\value{
A character vector.
}
\description{
Write the Stan code lines related with Density priors.
}
