% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{PriorOnlyDensity}
\alias{PriorOnlyDensity}
\title{Create a representation of a probability mass or density function for a
univariate random variable that may be used only to specify a prior
distribution for a model parameter.}
\usage{
PriorOnlyDensity(name, ...)
}
\arguments{
\item{name}{A character string with the name of the density.}

\item{...}{Other arguments for the density.}
}
\value{
A \code{\link{Density}} object.
}
\description{
It can be used to specify either a
prior distribution for a model parameter or a likelihood function for an
observation model.
}
\seealso{
Other PriorOnlyDensity: \code{\link{PriorOnlyMultivariateDensity}}
}
\concept{PriorOnlyDensity}
\keyword{internal}
