% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityRegBernoulliLogit.R
\name{RegBernoulliLogit}
\alias{RegBernoulliLogit}
\title{Bernoulli regression with logistic link density (univariate, discrete, binary space)}
\usage{
RegBernoulliLogit(xBeta = NULL, M = NULL, bounds = list(NULL, NULL),
  trunc = list(NULL, NULL), k = NULL, r = NULL, param = NULL)
}
\arguments{
\item{xBeta}{Either a fixed value or a prior density for the parameter of the regression.}

\item{M}{An integer with the number of covariates in the observation regression model.}

\item{bounds}{(optional) A list with two elements specifying the lower and upper bound for the parameter space. Use either a fixed value for a finite bound or NULL for no bounds. It defaults to an unbounded parameter space.}

\item{trunc}{(optional) A list with two elements specifying the lower and upper bound for the domain of the density function. Use either a fixed value for a finite bound or NULL for no truncation. It defaults to an unbounded domain.}

\item{k}{(optional) The number of the hidden state for which this density should be used. This argument is mostly for internal use: you should not use it unless you are acquainted with the internals of this software.}

\item{r}{(optional) The dimension of the observation vector dimension for which this density should be used. This argument is mostly for internal use: you should not use it unless you are acquainted with the internals of this software.}

\item{param}{(optional) The name of the parameter. This argument is mostly for internal use: you should not use it unless you are acquainted with the internals of this software.}
}
\value{
A \code{\link{Density}} object.
}
\description{
It can be used to specify either a
prior distribution for a model parameter or a likelihood function for an
observation model.
}
\examples{
RegBernoulliLogit(
  xBeta = Gaussian(0, 10),
  M     = 3
)
}
\seealso{
Other Density: \code{\link{Bernoulli}}, \code{\link{Beta}},
  \code{\link{Binomial}}, \code{\link{Categorical}},
  \code{\link{Cauchy}}, \code{\link{CholeskyLKJCor}},
  \code{\link{Density}}, \code{\link{Dirichlet}},
  \code{\link{Exponential}}, \code{\link{GammaDensity}},
  \code{\link{Gaussian}}, \code{\link{ImproperUniform}},
  \code{\link{InitialFixed}}, \code{\link{InitialSoftmax}},
  \code{\link{InverseWishart}},
  \code{\link{MVGaussianCholeskyCor}},
  \code{\link{MVGaussian}}, \code{\link{MVStudent}},
  \code{\link{Multinomial}},
  \code{\link{NegativeBinomialLocation}},
  \code{\link{NegativeBinomial}}, \code{\link{Poisson}},
  \code{\link{RegBinomialLogit}},
  \code{\link{RegBinomialProbit}},
  \code{\link{RegCategoricalSoftmax}},
  \code{\link{RegGaussian}}, \code{\link{Student}},
  \code{\link{TransitionFixed}},
  \code{\link{TransitionSoftmax}}, \code{\link{Wishart}}
}
\concept{Density}
