% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityRegBinomialLogit.R
\name{RegBinomialLogit}
\alias{RegBinomialLogit}
\title{Binomial regression with logistic link density (univariate, discrete, binary space)}
\usage{
RegBinomialLogit(xBeta = NULL, M = NULL, N = NULL,
  bounds = list(NULL, NULL), trunc = list(NULL, NULL), k = NULL,
  r = NULL, param = NULL)
}
\arguments{
\item{xBeta}{Either a fixed value or a prior density for the parameter of the regression.}

\item{M}{An integer with the number of covariates in the observation regression model.}

\item{N}{An integer with the number of trials (fixed quantity).}

\item{bounds}{(optional) A list with two elements specifying the lower and upper bound for the parameter space. Use either a fixed value for a finite bound or NULL for no bounds. It defaults to an unbounded parameter space.}

\item{trunc}{(optional) A list with two elements specifying the lower and upper bound for the domain of the density function. Use either a fixed value for a finite bound or NULL for no truncation. It defaults to an unbounded domain.}

\item{k}{(optional) The number of the hidden state for which this density should be used. This argument is mostly for internal use: you should not use it unless you are acquainted with the internals of this software.}

\item{r}{(optional) The dimension of the observation vector dimension for which this density should be used. This argument is mostly for internal use: you should not use it unless you are acquainted with the internals of this software.}

\item{param}{(optional) The name of the parameter. This argument is mostly for internal use: you should not use it unless you are acquainted with the internals of this software.}
}
\value{
#' @param x A \code{\link{Density}} object.
}
\description{
Binomial regression with logistic link density (univariate, discrete, binary space)
}
\examples{
RegBinomialLogit(
  xBeta = Gaussian(0, 10),
  M     = 3,
  N     = 10
)
}
\seealso{
Other Density: \code{\link{Binomial}},
  \code{\link{Categorical}}, \code{\link{Density}},
  \code{\link{InitialSoftmax}},
  \code{\link{RegBernoulliLogit}},
  \code{\link{RegBinomialProbit}},
  \code{\link{RegCategoricalSoftmax}},
  \code{\link{RegGaussian}}, \code{\link{TransitionFixed}},
  \code{\link{TransitionSoftmax}}
}
\concept{Density}
