% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{select_initial_parameters}
\alias{select_initial_parameters}
\title{Return the name of the initial model parameters.}
\usage{
select_initial_parameters(fit)
}
\arguments{
\item{fit}{An object returned by either \code{\link{draw_samples}} or \code{\link{optimizing}}.}
}
\value{
A character vector with the names of the initial model parameters.
}
\description{
Return the name of the initial model parameters.
}
\seealso{
See \code{\link{select_parameters}}.
}
\keyword{internal}
