% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{select_parameters}
\alias{select_parameters}
\title{Return the name of the model parameters.}
\usage{
select_parameters(fit, observation = TRUE, initial = TRUE,
  transition = TRUE)
}
\arguments{
\item{fit}{An object returned by either \code{\link{sampling}} or \code{\link{optimizing}}.}

\item{observation}{An optional logical scalar indicating whether the names of the observation model parameters should be included. Defaults to TRUE.}

\item{initial}{An optional logical scalar indicating whether the names of the initial model parameters should be included. Defaults to TRUE.}

\item{transition}{An optional logical scalar indicating whether the names of the transition model parameters should be included. Defaults to TRUE.}
}
\value{
A character vector with the selected model parameter names.
}
\description{
Return the name of the model parameters.
}
\keyword{internal}
