% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{select_transition_parameters}
\alias{select_transition_parameters}
\title{Return the name of the transition model parameters.}
\usage{
select_transition_parameters(fit)
}
\arguments{
\item{fit}{An object returned by either \code{\link{sampling}} or \code{\link{optimizing}}.}
}
\value{
A character vector with the names of the transition model parameters.
}
\description{
Return the name of the transition model parameters.
}
\seealso{
\code{\link{select_parameters}}.
}
\keyword{internal}
