% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{sim}
\alias{sim}
\title{Simulate data from the prior predictive density.}
\usage{
sim(spec, T = 1000, x = NULL, u = NULL, v = NULL, nSimulations = 500, ...)
}
\arguments{
\item{spec}{An object returned by either \code{\link{specify}} or \code{\link{hmm}}.}

\item{...}{Arguments to be passed to rstan's \code{\link[rstan]{sampling}}.}

\item{nSimulations}{An optional integer with the number of simulations. It defaults to 500 time series.}
}
\description{
Simulate data from the prior predictive density.
}
