% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{spec}
\alias{spec}
\title{Create a model specification}
\usage{
spec(K, R, observation = NULL, initial = NULL, transition = NULL,
  name = "")
}
\arguments{
\item{K}{An integer with the number of hidden states.}

\item{R}{An integer with the dimension of the observation vector (e.g. one is univariate, two is bivariate)}

\item{observation}{One density, or more than one density chained with the `+` operator, describing the observation model. See below and VIGNETTE for detailed explanation.}

\item{initial}{One density, or more than one density chained with the `+` operator, describing the initial distribution model. See below and VIGNETTE for detailed explanation.}

\item{transition}{One density, or more than one density chained with the `+` operator, describing the transition model. See below and VIGNETTE for detailed explanation.}

\item{name}{An optional string with a name for a model.}
}
\value{
An specification object that may be used to generate data from or fit a model.
}
\description{
Create a model specification
}
\section{Model specification}{

Explanation of model specification here.
}

