% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortChain.R
\name{stan_sort_chain}
\alias{stan_sort_chain}
\title{Sorts the samples from a stanfit object.}
\usage{
stan_sort_chain(stanfit, reference, K)
}
\description{
It's very experimental stuff to fix label switching - BEWARE
}
\keyword{internals}
