% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{string_simplify}
\alias{string_simplify}
\title{Remove non-alphabetic characters and modify to lowercase.}
\usage{
string_simplify(string)
}
\arguments{
\item{string}{A character string, or a vector of character strings}
}
\value{
A character string, or a vector of character strings.
}
\description{
Remove non-alphabetic characters and modify to lowercase.
}
