% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme}
\alias{theme}
\title{Theme for BayesHMM visualizations and printouts}
\description{
BayesHMM comes with many built-in visualization and text printout methods to carry out bayesian data analysis out of the box. All plot colors and printout text formats are based on a fully-customizable theme loaded as a global setting. The theme is simply a named list with two elements \emph{BayesHMM.theme} and \emph{BayesHMM.print}, which in turn are named lists following the structure described below. Run \code{\link{get_default_theme}} to see the default values.
}
\details{
\strong{BayesHMM.theme}: settings for visualizations.
\describe{
  \item{boxY}{A string with the color of the lines of the box corresponding to the observed sample. Applies to \code{\link{plot_ppredictive}} with features \emph{boxplot}.}
  \item{boxYPred}{A string with the color of the lines of the many boxes corresponding to the samples drawn from the posterior predictive density. Applies to \code{\link{plot_ppredictive}} with features \emph{boxplot}.}
  \item{cumulativeY}{A string with the color of the cumulative density line for the observed sample. Applies to \code{\link{plot_ppredictive}} with features \emph{cumulative}.}
  \item{cumulativeYPred}{A string with the color of the cumulative density lines for the many samples drawn from the posterior predictive density. Applies to \code{\link{plot_ppredictive}} with features \emph{cumulative}.}
  \item{densityY}{A string with the color of the density line for the observed sample. Applies to \code{\link{plot_ppredictive}} with features \emph{density}.}
  \item{densityYPred}{A string with the color of the density lines for the many samples drawn from the posterior predictive density. Applies to \code{\link{plot_ppredictive}} with features \emph{density}.}
  \item{histBorder}{A string with the color of the border of the histogram bars. Applies to \code{\link{plot_ppredictive}} with features \emph{histogram}.}
  \item{histCol}{A string with the color of the area of the histogram bars. Applies to \code{\link{plot_ppredictive}} with features \emph{histogram}.}
  \item{histLine}{A string with the color of the vertical line related to the \emph{fun} argument of the \code{\link{plot_ppredictive}}) function. The line is located at the value returned by \emph{fun} applied to the actual sample. Applies to \code{\link{plot_ppredictive}} with features \emph{histogram}.}
  \item{ksBorder}{A string with the color of the border of the histogram bars. Applies to \code{\link{plot_ppredictive}} with features \emph{ks}.}
  \item{ksCol}{A string with the color of the area of the histogram bars. Applies to \code{\link{plot_ppredictive}} with features \emph{ks}.}
  \item{scatterY}{A string with the color of the point for the observed sample. Applies to \code{\link{plot_ppredictive}} with features \emph{scatterplot}.}
  \item{scatterYPred}{A string with the color of the points for the many samples drawn from the posterior predictive density. Applies to \code{\link{plot_ppredictive}} with features \emph{scatterplot}.}
  \item{seriesY}{A string with the color of the observation time series. Applies to \code{\link{plot_series}} and \code{\link{plot_state_probability}}.}
  \item{shadeAlpha}{A numeric value between zero and one with the transparency setting for the shaded features, where lower values lead to more transparent shading. Applies to \code{\link{plot_series}} and \code{\link{plot_state_probability}}.}
  \item{states}{A vector of strings with the colors of each latent state. Applies to \code{\link{plot_series}} and \code{\link{plot_state_probability}}.}
}

Note that the overall legend uses the colors \emph{densityY} and \emph{densityYPred} for the actual sample and the posterior predictive samples respectively.

\strong{BayesHMM.print}: settings for text printouts.
\describe{
  \item{char}{A character to use to print text lines and boxes. It defaults to an underscore.}
  \item{tab}{A string to be used for indentation. It defaults to two spaces.}
  \item{textWidth}{An integer with the width of the text lines and boxes. It defaults to 80 characters.}
}
}
\seealso{
Other visualization functions: \code{\link{get_current_theme}},
  \code{\link{get_default_theme}},
  \code{\link{plot_ppredictive}},
  \code{\link{plot_series}},
  \code{\link{plot_state_probability}}
}
\concept{visualization functions}
