% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityTransitionFixed.R
\name{TransitionFixed}
\alias{TransitionFixed}
\title{Fixed transition matrix}
\usage{
TransitionFixed(A = NULL, bounds = list(NULL, NULL),
  trunc = list(NULL, NULL), k = NULL, r = NULL, param = NULL)
}
\arguments{
\item{A}{A numeric matrix of size \emph{KxK} with rows suming to one.}

\item{bounds}{(optional) A list with two elements specifying the lower and upper bound for the parameter space. Use either a fixed value for a finite bound or NULL for no bounds. It defaults to an unbounded parameter space.}

\item{trunc}{(optional) A list with two elements specifying the lower and upper bound for the domain of the density function. Use either a fixed value for a finite bound or NULL for no truncation. It defaults to an unbounded domain.}

\item{k}{(optional) The number of the hidden state for which this density should be used. This argument is mostly for internal use: you should not use it unless you are acquainted with the internals of this software.}

\item{r}{(optional) The dimension of the observation vector dimension for which this density should be used. This argument is mostly for internal use: you should not use it unless you are acquainted with the internals of this software.}

\item{param}{(optional) The name of the parameter. This argument is mostly for internal use: you should not use it unless you are acquainted with the internals of this software.}
}
\value{
#' @param x A \code{\link{Density}} object.
}
\description{
Fixed transition matrix
}
\examples{
TransitionFixed(
  A = matrix(c(0.2, 0.8, 0.9, 0.1), 2, 2)
)
}
\seealso{
Other Density: \code{\link{Binomial}},
  \code{\link{Categorical}}, \code{\link{Density}},
  \code{\link{InitialSoftmax}},
  \code{\link{RegBernoulliLogit}},
  \code{\link{RegBinomialLogit}},
  \code{\link{RegBinomialProbit}},
  \code{\link{RegCategoricalSoftmax}},
  \code{\link{RegGaussian}},
  \code{\link{TransitionSoftmax}}
}
\concept{Density}
