% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityWishart.R
\name{Wishart}
\alias{Wishart}
\title{Wishart density (multivariate, continuous, bounded space, prior only)}
\usage{
Wishart(nu = NULL, sigma = NULL, bounds = list(NULL, NULL),
  trunc = list(NULL, NULL), k = NULL, r = NULL, param = NULL)
}
\arguments{
\item{nu}{Either a fixed value or a prior density for the degree-of-freedom scalar parameter.}

\item{sigma}{Either a fixed value or a prior density for the scale matrix parameter.}

\item{bounds}{(optional) A list with two elements specifying the lower and upper bound for the parameter space. Use either a fixed value for a finite bound or NULL for no bounds. It defaults to an unbounded parameter space.}

\item{trunc}{(optional) A list with two elements specifying the lower and upper bound for the domain of the density function. Use either a fixed value for a finite bound or NULL for no truncation. It defaults to an unbounded domain.}

\item{k}{(optional) The number of the hidden state for which this density should be used. This argument is mostly for internal use: you should not use it unless you are acquainted with the internals of this software.}

\item{r}{(optional) The dimension of the observation vector dimension for which this density should be used. This argument is mostly for internal use: you should not use it unless you are acquainted with the internals of this software.}

\item{param}{(optional) The name of the parameter. This argument is mostly for internal use: you should not use it unless you are acquainted with the internals of this software.}
}
\value{
A Density object.
}
\description{
Wishart density (multivariate, continuous, bounded space, prior only)
}
\examples{
# As a prior for the precision matrix
Wishart(nu = 5, sigma = matrix(c(1, 0, 0, 1), 2, 2))
}
\seealso{
Other Density
#': \code{\link{Bernoulli}},
  \code{\link{Beta}}, \code{\link{Binomial}},
  \code{\link{Categorical}}, \code{\link{Cauchy}},
  \code{\link{CholeskyLKJCor}}, \code{\link{Default}},
  \code{\link{Dirichlet}}, \code{\link{Exponential}},
  \code{\link{GammaDensity}}, \code{\link{Gaussian}},
  \code{\link{InverseWishart}},
  \code{\link{MVGaussianCholeskyCor}},
  \code{\link{MVGaussian}}, \code{\link{MVStudent}},
  \code{\link{Multinomial}},
  \code{\link{NegativeBinomialLocation}},
  \code{\link{NegativeBinomial}}, \code{\link{Poisson}},
  \code{\link{Student}}
}
\concept{Density
#'}
