% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitIntegration.R
\docType{methods}
\name{browse_model}
\alias{browse_model}
\alias{browse_model.Optimization}
\alias{browse_model,stanfit-method}
\title{Load the underlying Stan code into an IDE or browser (platform dependent).}
\usage{
browse_model(fit, ...)

## S3 method for signature 'Optimization'
browse_model(fit, ...)
#'

## S4 method for signature 'stanfit'
browse_model(fit, ...)
#'
}
\arguments{
\item{fit}{An object returned by either \code{\link{drawSamples}} or \code{\link{optimizing}}.}

\item{...}{(optional) Arguments to be passed to \code{\link{browseURL}} (namely \emph{browser} and \emph{encodeIfNeeded}).}
}
\value{
No return value.
}
\description{
Load the underlying Stan code into an IDE or browser (platform dependent).
}
\seealso{
\code{\link{browseURL}}.
#'
}
