% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{browse_model.Specification}
\alias{browse_model.Specification}
\title{Load the underlying Stan code into an IDE or browser (platform dependent).}
\usage{
browse_model.Specification(spec)
}
\arguments{
\item{spec}{An object returned by either \code{\link{specify}} or \code{\link{hmm}}.}
}
\value{
No return value.
}
\description{
Load the underlying Stan code into an IDE or browser (platform dependent).
}
\seealso{
\code{\link{browseURL}}.
}
