% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityInitialSoftmax.R
\name{explain_density.InitialSoftmax}
\alias{explain_density.InitialSoftmax}
\title{Explain a Density object in human readable format.}
\usage{
\method{explain_density}{InitialSoftmax}(x, print = TRUE)
}
\arguments{
\item{x}{A \code{\link{Density}} object.}

\item{print}{An optional logical indicating whether the description should be printing out.}
}
\value{
A character string.
}
\description{
Explain a Density object in human readable format.
}
\keyword{internal}
