% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\docType{methods}
\name{extract_date}
\alias{extract_date}
\alias{extract_date.Optimization}
\alias{extract_date,stanfit-method}
\title{Extract the date when the model was run.}
\usage{
extract_date(fit)

## S3 method for signature 'Optimization'
extract_date(fit)

## S4 method for signature 'stanfit'
extract_date(fit)
}
\arguments{
\item{fit}{An object returned by either \code{\link{draw_samples}} or \code{\link{optimizing}}.}
}
\value{
An integer with the date used to fit the model.
}
\description{
Extract the date when the model was run.
}
