% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfit.R
\docType{methods}
\name{extract_n_chains}
\alias{extract_n_chains}
\alias{extract_n_chains,stanfit-method}
\title{Extract the number of chains (M).}
\usage{
extract_n_chains(stanfit)

\S4method{extract_n_chains}{stanfit}(stanfit)
}
\arguments{
\item{stanfit}{An object returned by either \code{\link{fit}} or \code{\link{drawSamples}}.}
}
\value{
The number of chains (M) used to fit the model.
#'Method extract_n_chains
}
\description{
Extract the number of chains (M).
}
