% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfit.R
\name{extract_n_iterations}
\alias{extract_n_iterations}
\title{Extract the number of total iterations.}
\usage{
extract_n_iterations(stanfit)
}
\arguments{
\item{stanfit}{An object returned by either \code{\link{fit}} or \code{\link{draw_samples}}.}
}
\value{
The number of total iterations.
}
\description{
Extract the number of total iterations.
}
