% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfit.R
\docType{methods}
\name{extract_n_thin}
\alias{extract_n_thin}
\alias{extract_n_thin,stanfit-method}
\title{Extract the thinning periodicity.}
\usage{
extract_n_thin(stanfit)

\S4method{extract_n_thin}{stanfit}(stanfit)
}
\arguments{
\item{stanfit}{An object returned by either \code{\link{fit}} or \code{\link{draw_samples}}.}
}
\value{
The the thinning periodicity.
}
\description{
Extract the thinning periodicity.
}
