% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\docType{methods}
\name{extract_parameter_names}
\alias{extract_parameter_names}
\alias{extract_parameter_names.Optimization}
\alias{extract_parameter_names,stanfit-method}
\title{Extract the names of all the model parameters in the fit object.}
\usage{
extract_parameter_names(fit)

## S3 method for signature 'Optimization'
extract_parameter_names(fit)

## S4 method for signature 'stanfit'
extract_parameter_names(fit)
}
\arguments{
\item{fit}{An object returned by either \code{\link{draw_samples}} or \code{\link{optimizing}}.}
}
\value{
A character vector with the selected model parameter names.
}
\description{
Extract the names of all the model parameters in the fit object.
}
\seealso{
See \code{\link{select_parameters}} for extracting the names of part of the model.
}
\keyword{internal}
