% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{extract_seed}
\alias{extract_seed}
\title{Extract the time elapsed when fitting the model.}
\usage{
extract_seed(fit)
}
\arguments{
\item{fit}{An object returned by either \code{\link{sampling}} or \code{\link{optimizing}}.}
}
\value{
An integer with the seed used to fit the model.
}
\description{
Extract the time elapsed when fitting the model.
}
