% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\docType{methods}
\name{extract_seed}
\alias{extract_seed}
\alias{extract_seed.Optimization}
\alias{extract_seed,stanfit-method}
\title{Extract the time elapsed when fitting the model.}
\usage{
extract_seed(fit)

## S3 method for signature 'Optimization'
extract_seed(fit)

## S4 method for signature 'stanfit'
extract_seed(fit)
}
\arguments{
\item{fit}{An object returned by either \code{\link{draw_samples}} or \code{\link{optimizing}}.}
}
\value{
An integer with the seed used to fit the model.
}
\description{
Extract the time elapsed when fitting the model.
}
