% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{extract_stanmodel.stanmodel}
\alias{extract_stanmodel.stanmodel}
\title{Extract the stanmodel object of the fitted object.}
\usage{
\method{extract_stanmodel}{stanmodel}(fit)
}
\arguments{
\item{fit}{An object returned by either \code{\link{draw_samples}} or \code{\link{optimizing}}.}
}
\value{
An integer with the stanmodel used to fit the model.
}
\description{
Extract the stanmodel object of the fitted object.
}
\keyword{internal}
