% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitIntegration.R
\docType{methods}
\name{extract_time}
\alias{extract_time}
\alias{extract_time.Optimization}
\alias{extract_time,stanfit-method}
\title{Extract the time elapsed when fitting the model.}
\usage{
extract_time(fit)

## S3 method for signature 'Optimization'
extract_time(fit)

## S4 method for signature 'stanfit'
extract_time(fit)
}
\arguments{
\item{fit}{An object returned by either \code{\link{sampling}} or \code{\link{optimizing}}.}
}
\value{
For objects fit via \code{\link{sampling}}, a named numeric matrix with two columns (warm-up and sample time) and one row per chain. For objects fit via \code{\link{optimizing}}, the system time elapsed.
}
\description{
Extract the time elapsed when fitting the model.
}
