% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityCategorical.R
\name{fixedParameters.Categorical}
\alias{fixedParameters.Categorical}
\title{Write the Stan code lines related with Density fixed parameters.}
\usage{
\method{fixedParameters}{Categorical}(x)
}
\arguments{
\item{x}{A \code{\link{Density}} object.}
}
\value{
A character vector.
}
\description{
Write the Stan code lines related with Density fixed parameters.
}
