% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{is.Density}
\alias{is.Density}
\title{Check if it is a \code{\link{Density}} object.}
\usage{
is.Density(x)
}
\arguments{
\item{x}{An object.}
}
\value{
TRUE if the object is a Density object, FALSE otherwise.
}
\description{
Check if it is a \code{\link{Density}} object.
}
\keyword{internal}
