% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{is.DensityList}
\alias{is.DensityList}
\title{Check if it is a Density object.}
\usage{
is.DensityList(x)
}
\arguments{
\item{x}{An object.}
}
\value{
TRUE if the object is a Density objectm FALSE otherwise.
}
\description{
Check if it is a Density object.
}
\examples{
is.Density(
  Gaussian(0, 1)
)

is.Density(1 + 1)
}
