% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{is.link}
\alias{is.link}
\title{Check if it is a Density object for the transition model.}
\usage{
is.link(x)
}
\arguments{
\item{x}{a Density object.}
}
\value{
TRUE if the object is meant for transition models, FALSE otherwise.
}
\description{
Check if it is a Density object for the transition model.
}
\examples{
is.link(
  TransitionSoftmax(uBeta = Gaussian(0, 10), P = 2)
)
}
