% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{+.Density}
\alias{+.Density}
\alias{+}
\alias{DensityList}
\title{Append two Density objects.}
\usage{
x + y
}
\arguments{
\item{x}{A Density object (e.g. \code{\link{Gaussian}})}

\item{y}{A Density object (e.g. \code{\link{Gaussian}})}
}
\value{
A DensityList object.
}
\description{
Append two Density objects.
}
\examples{
Gaussian(0, 1) + Gaussian(0, 1)
}
