% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{PriorOnlyMultivariateDensity}
\alias{PriorOnlyMultivariateDensity}
\title{Create a representation of a probability mass or density function for a
multivariate random variable that may be used only to specify a prior
distribution for a model parameter.}
\usage{
PriorOnlyMultivariateDensity(name, ...)
}
\arguments{
\item{name}{A character string with the name of the density.}

\item{...}{Other arguments for the density.}
}
\value{
A \code{\link{Density}} object.
}
\description{
It can be used to specify either a
prior distribution for a model parameter or a likelihood function for an
observation model.
}
\references{
Betancourt, Michael (2017) Identifying Bayesian Mixture Models \emph{Stan Case Studies} \bold{Volume 4}. \href{https://mc-stan.org/users/documentation/case-studies/identifying_mixture_models.html}{Link}.
}
\seealso{
Other PriorOnlyDensity: \code{\link{PriorOnlyDensity}}
}
\concept{PriorOnlyDensity}
\keyword{internal}
