% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{run}
\alias{run}
\title{Run a Markov-chain Monte Carlo algorithm to sample from the log posterior density.}
\usage{
run(spec, data = NULL, writeDir = tempdir(), ...)
}
\arguments{
\item{spec}{An object returned by either \code{\link{specify}} or \code{\link{hmm}}.}

\item{writeDir}{An optional string with the path where the Stan file should be written. Useful to inspect and modify the Stan code manually. It defaults to a temporary directory.}

\item{...}{Arguments to be passed to rstan's \code{\link[rstan]{stan}}.}
}
\value{
An object of S4 class stanfit with some additional attributes (the dataset \emph{data}, the name of the Stan code file \emph{filename}, and the specification object \emph{spec}). This object is completely compatible with all other functions.
}
\description{
Run a Markov-chain Monte Carlo algorithm to sample from the log posterior density.
}
\seealso{
See rstan's \code{\link[rstan]{stan}} and \code{\link[rstan]{sampling}} for further details on tunning the MCMC algorithm.
}
\keyword{internal}
