% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{select_all_parameters}
\alias{select_all_parameters}
\title{Return the name of all model parameters.}
\usage{
select_all_parameters(fit)
}
\arguments{
\item{fit}{An object returned by either \code{\link{draw_samples}} or \code{\link{optimizing}}.}
}
\value{
A character vector with the names of all model parameters.
}
\description{
Return the name of all model parameters.
}
\seealso{
\code{\link{select_parameters}}.
}
\keyword{internal}
