% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R, R/fit.R
\name{select_all_parameters}
\alias{select_all_parameters}
\title{Return the name of all model parameters.}
\usage{
select_all_parameters(fit)

select_all_parameters(fit)
}
\arguments{
\item{fit}{An object returned by either \code{\link{sampling}} or \code{\link{optimizing}}.}

\item{fit}{An object returned by either \code{\link{sampling}} or \code{\link{optimizing}}.}
}
\value{
A character vector with the names of all model parameters.

A character vector with the names of all model parameters.
}
\description{
Return the name of all model parameters.

Return the name of all model parameters.
}
\seealso{
\code{\link{select_parameters}}.

\code{\link{select_parameters}}.
}
\keyword{internal}
